requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
drf-yasg
eqator
sentry-sdk
ultralytics
